Update date format in profile card component

+++ client/app/(profiles)/profiles/components/Hero/Profiles/Card.jsx
@@ -242,7 +242,7 @@ export default function Card(props) {
                     !haveCustomColors ? 'text-primary' : `text-[${variables.textPrimary}]`
                   )}
                 >
-                  {new Date(props.createdAt).toLocaleDateString('en-US', { year: 'numeric', month: 'long', day: 'numeric' })}
+                  {new Date(props.createdAt).toLocaleDateString('en-US', { year: 'numeric', month: 'short', day: 'numeric' })}
                 </p>
               </div>
             </div>
